feat: move a11y, bidi, platform, and portal to cdk
