+++ src/cdk/public_api.ts
@@ -6,5 +6,11 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+export * from './a11y/index';
+export * from './bidi/index';
 export * from './coercion/index';
 export * from './table/index';
+export * from './platform/index';
+export * from './portal/index';
+export * from './rxjs/index';
+export * from './keyboard/keycodes';
