+++ src/lib/core/bidi/index.ts
@@ -6,26 +6,12 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {NgModule} from '@angular/core';
-import {DOCUMENT} from '@angular/platform-browser';
-import {Dir} from './dir';
-import {DIR_DOCUMENT, Directionality, DIRECTIONALITY_PROVIDER} from './directionality';
-
 
 export {
   Directionality,
   DIRECTIONALITY_PROVIDER,
   DIR_DOCUMENT,
   Direction,
-} from './directionality';
-export {Dir} from './dir';
-
-@NgModule({
-  exports: [Dir],
-  declarations: [Dir],
-  providers: [
-    {provide: DIR_DOCUMENT, useExisting: DOCUMENT},
-    Directionality,
-  ]
-})
-export class BidiModule { }
+  Dir,
+  BidiModule,
+} from '@angular/cdk';
