+++ src/lib/core/platform/index.ts
@@ -6,15 +6,6 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {NgModule} from '@angular/core';
-import {Platform} from './platform';
-
-
-@NgModule({
-  providers: [Platform]
-})
-export class PlatformModule {}
-
-
+export {PlatformModule} from '@angular/cdk';
 export * from './platform';
 export * from './features';
